Revert "#1057 tools that modify FGs refer canvas correctly (#1064)"
This reverts commit 8e9e05e8b79d1d0da6774c7293121ea6f3e499a3.
